utils/auth_utils.py
from datetime import timedelta, datetime
from jose import JWTError, jwt
from fastapi.security import OAuth2PasswordBearer
from sqlalchemy.orm import Session

from models import UserModel

JWT_SECRET = "secret"
JWT_ALGORITHM = "HS256"

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/users/login")


def authenticate_user(db: Session, email: str, password: str):
    user = db.query(UserModel).filter(UserModel.email == email).first()
    if not user or not user.verify_password(password):
        return False

    # remove password from user object
    delattr(user, 'password')

    return user


def create_access_token(user: UserModel, expires_delta: timedelta | None = None):
    to_encode = {
        "email": user.email,
        "username": user.username,
        "id": user.id,
    }
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, JWT_SECRET, algorithm=JWT_ALGORITHM)
