requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-multipart
python-jose[cryptography]